app/backend/requirements.txt
boto3
psycopg2-binary
uvicorn
fastapi[standard]
typing
pydantic
minio